feat: refactor address/contract Component filter and contract
